working : 재계산쪽
