+++ src/js/priceStepFirst.js
